+++ HISTORY.rst
@@ -12,17 +12,18 @@ History
     * for all ResourcesT like :class:`tickets.Tickets` ``.url`` is now a property
     * exchange ``tickets.Ticket.merge_with()`` with :meth:`tickets.Ticket.merge_into()`
       mimicking the logic provided by the Web UI
-    * :meth:`tags.Tags.delete()` and :meth:`tags.Tags.delete()` now only accept the tag name
+    * :meth:`tags.Tags.delete()` and :meth:`tags.Tags.rename()` now only accept the tag name
     * changed signature of :meth:`tickets.Tickets.create()` to assure an article body or the article itself is provided
     * remove :attr:`articles.Article.encoding` property
 
 * **added features**
 
-    * extend supported Python version to 3.8
+    * extend supported Python version including Python 3.8
     * added :meth:`tickets.Ticket.history` method
     * added ``weburl`` property for :class:`client.Client`, :class:`tickets.Ticket`,
       :class:`users.User` and :class:`organizations.Organization`
     * added property :attr:`tickets.Ticket.create_article_sender`
+    * added property :attr:`groups.Group.parent_group`
 
 * **fixes**