+++ HISTORY.rst
@@ -25,6 +25,7 @@ History
     * added property :attr:`tickets.Ticket.create_article_sender`
     * added property :attr:`groups.Group.parent_group`
     * save timestamp when cache is updated, added method :meth:`tickets.Ticket.last_request_at`
+    * ``dir(Resource)`` now also returns the dynamic attributes
 
 * **fixes**